+++ crates/rattler_networking/src/lib.rs
@@ -3,11 +3,13 @@
 //! Networking utilities for Rattler, specifically authenticating requests
 pub use authentication_middleware::AuthenticationMiddleware;
 pub use authentication_storage::{authentication::Authentication, storage::AuthenticationStorage};
-pub use mirror_middleware::{MirrorMiddleware, OciMiddleware};
+pub use mirror_middleware::MirrorMiddleware;
+pub use oci_middleware::OciMiddleware;
 
 pub mod authentication_middleware;
 pub mod authentication_storage;
 pub mod mirror_middleware;
+pub mod oci_middleware;
 pub mod retry_policies;
 
 mod redaction;
